+++ packages/dumb/src/components/settings/history/history-table-content.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,44 +25,33 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import { Auth0Provider } from '@auth0/auth0-react'
+import { HistorySpannedRow } from './history-table-spanned-row'
+import { HistoryTableRow } from './history-table-row'
+import React, { type FunctionComponent } from 'react'
+import { type TimePrefs } from 'medical-domain'
+import { type ChangeDateParameterGroup } from '../../../models/historized-parameter.model'
+import { transformToHistorizedParameters } from './history-table.core'
+import type moment from 'moment-timezone'
 
-import '@fontsource/roboto/300.css'
-import '@fontsource/roboto/400.css'
-import '@fontsource/roboto/500.css'
-import '@fontsource/roboto/700.css'
-import 'branding/theme.css'
-import 'classes.css'
-
-import appConfig from '../lib/config/config'
-import { AuthContextProvider } from '../lib/auth'
-import { MainLobby } from './main-lobby'
-import MetricsLocationListener from '../components/MetricsLocationListener'
-
-const Yourloops = (): JSX.Element => {
-  const redirectUri = window.location.origin
-  return (
-    <Auth0Provider
-      domain={appConfig.AUTH0_DOMAIN}
-      issuer={appConfig.AUTH0_ISSUER}
-      clientId={appConfig.AUTH0_CLIENT_ID}
-      useRefreshTokensFallback
-      authorizationParams={{
-        redirectUri,
-        audience: 'https://api-ext.your-loops.com'
-      }}
-      useRefreshTokens
-    >
-      <BrowserRouter>
-        <MetricsLocationListener />
-        <AuthContextProvider>
-          <MainLobby />
-        </AuthContextProvider>
-      </BrowserRouter>
-    </Auth0Provider>
-  )
+interface HistoryTableContentProps {
+  onSwitchToDaily: (date: moment.Moment | Date | number | null) => void
+  rows: ChangeDateParameterGroup[]
+  timePrefs: TimePrefs
 }
 
-export default Yourloops
+export const HistoryTableContent: FunctionComponent<HistoryTableContentProps> = (props): JSX.Element => {
+  const { onSwitchToDaily, rows, timePrefs } = props
+  const historizedParameters = transformToHistorizedParameters(rows, timePrefs)
+
+  return (
+    <tbody>
+    {
+      historizedParameters.map((row, key) =>
+        row.isGroupedParameterHeader
+          ? (<HistorySpannedRow key={key} data={row} onSwitchToDaily={onSwitchToDaily} />)
+          : (<HistoryTableRow key={key} data={row} />)
+      )
+    }
+    </tbody>
+  )
+}