+++ packages/dumb/src/components/settings/history/history-table-header.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,44 +25,28 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import { Auth0Provider } from '@auth0/auth0-react'
+import React, { type FunctionComponent } from 'react'
+import styles from '../diabeloop.css'
+import { useTranslation } from 'react-i18next'
 
-import '@fontsource/roboto/300.css'
-import '@fontsource/roboto/400.css'
-import '@fontsource/roboto/500.css'
-import '@fontsource/roboto/700.css'
-import 'branding/theme.css'
-import 'classes.css'
-
-import appConfig from '../lib/config/config'
-import { AuthContextProvider } from '../lib/auth'
-import { MainLobby } from './main-lobby'
-import MetricsLocationListener from '../components/MetricsLocationListener'
-
-const Yourloops = (): JSX.Element => {
-  const redirectUri = window.location.origin
+export const HistoryTableHeader: FunctionComponent = () => {
+  const { t } = useTranslation('main')
   return (
-    <Auth0Provider
-      domain={appConfig.AUTH0_DOMAIN}
-      issuer={appConfig.AUTH0_ISSUER}
-      clientId={appConfig.AUTH0_CLIENT_ID}
-      useRefreshTokensFallback
-      authorizationParams={{
-        redirectUri,
-        audience: 'https://api-ext.your-loops.com'
-      }}
-      useRefreshTokens
-    >
-      <BrowserRouter>
-        <MetricsLocationListener />
-        <AuthContextProvider>
-          <MainLobby />
-        </AuthContextProvider>
-      </BrowserRouter>
-    </Auth0Provider>
+    <thead>
+      <tr>
+        <th className={styles.secondaryLabelAlone}>
+          {t('Level')}
+        </th>
+        <th className={styles.secondaryLabelAlone}>
+          {t('Parameter')}
+        </th>
+        <th className={styles.secondaryLabelAlone}>
+          {t('Value')}
+        </th>
+        <th className={styles.secondaryLabelAlone}>
+          {t('Date')}
+        </th>
+      </tr>
+    </thead>
   )
 }
-
-export default Yourloops