+++ packages/dumb/src/components/settings/history/history-table.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,44 +25,33 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react'
-import { BrowserRouter } from 'react-router-dom'
-import { Auth0Provider } from '@auth0/auth0-react'
+import React, { type FunctionComponent } from 'react'
+import styles from '../diabeloop.css'
+import { type TimePrefs } from 'medical-domain'
+import { useTranslation } from 'react-i18next'
+import { HistoryTableHeader } from './history-table-header'
+import { HistoryTableContent } from './history-table-content'
+import { type ChangeDateParameterGroup } from '../../../models/historized-parameter.model'
+import type moment from 'moment-timezone'
 
-import '@fontsource/roboto/300.css'
-import '@fontsource/roboto/400.css'
-import '@fontsource/roboto/500.css'
-import '@fontsource/roboto/700.css'
-import 'branding/theme.css'
-import 'classes.css'
-
-import appConfig from '../lib/config/config'
-import { AuthContextProvider } from '../lib/auth'
-import { MainLobby } from './main-lobby'
-import MetricsLocationListener from '../components/MetricsLocationListener'
-
-const Yourloops = (): JSX.Element => {
-  const redirectUri = window.location.origin
-  return (
-    <Auth0Provider
-      domain={appConfig.AUTH0_DOMAIN}
-      issuer={appConfig.AUTH0_ISSUER}
-      clientId={appConfig.AUTH0_CLIENT_ID}
-      useRefreshTokensFallback
-      authorizationParams={{
-        redirectUri,
-        audience: 'https://api-ext.your-loops.com'
-      }}
-      useRefreshTokens
-    >
-      <BrowserRouter>
-        <MetricsLocationListener />
-        <AuthContextProvider>
-          <MainLobby />
-        </AuthContextProvider>
-      </BrowserRouter>
-    </Auth0Provider>
-  )
+interface HistoryParameterTableProps {
+  rows: ChangeDateParameterGroup[]
+  onSwitchToDaily: (date: moment.Moment | Date | number | null) => void
+  timePrefs: TimePrefs
 }
 
-export default Yourloops
+export const HistoryParameterTable: FunctionComponent<HistoryParameterTableProps> = (props) => {
+  const { t } = useTranslation('main')
+  const { rows, onSwitchToDaily, timePrefs } = props
+
+  return (
+    <table className={styles.settingsTable}>
+      <caption className={styles.bdlgSettingsHeader}>
+        {t('Parameters History')}
+        <span className={styles.secondaryLabelWithMain} />
+      </caption>
+      <HistoryTableHeader />
+      <HistoryTableContent rows={rows} onSwitchToDaily={onSwitchToDaily} timePrefs={timePrefs} />
+    </table>
+  )
+}