+++ CHANGELOG.md
@@ -16,6 +16,7 @@ It is based on Tidepool Blip 1.27.
 ### Engineering use
 - YLP-2044 Remove redux settings store
 - YLP-2065 Migrate CBGDateTraceLabel from viz to dumb
+- YLP-2066 Migrate historized settings table
 - YLP-2068 Update minor dependencies
 - YLP-2069 Migrate CBGSlicesContainer from viz to dumb
 - YLP-2071 Change the question mark icon to accept button
@@ -47,6 +48,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2196 Removed the team selection drop-down menu on the care team definition page
 - YLP-2217 Bump react-router-dom to 6.8.2
 - YLP-2219 Migrate SensorUsage from stats.js to yourloops
+- YLP-2252 Bump auth0-react from 1.9.0 to 2.0.1
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team